new_traffic_god/requirements.txt
# New Traffic God - Requirements
# ================================

# Core ML/DL
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Transformer / NLP
transformers>=4.30.0
tokenizers>=0.13.0
sentencepiece>=0.1.99

# Vector Store / RAG
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support
sentence-transformers>=2.2.0
rank-bm25>=0.2.2

# API
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0
polars>=0.18.0  # Faster DataFrame operations
pyarrow>=12.0.0

# Geospatial
geopy>=2.3.0
shapely>=2.0.0
osmnx>=1.3.0  # OpenStreetMap data

# Visualization (optional, for analysis)
matplotlib>=3.7.0
plotly>=5.14.0
folium>=0.14.0  # Map visualization

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Async
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0  # Better terminal output
loguru>=0.7.0  # Better logging

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
ruff>=0.0.270

# Optional: GPU acceleration
# torch-cuda  # Uncomment for CUDA support
# cupy>=12.0.0  # GPU arrays

# Optional: Distributed training
# deepspeed>=0.9.0
# accelerate>=0.20.0

# Optional: Model optimization
# onnx>=1.14.0
# onnxruntime>=1.15.0
